refactor(tests): replace inline `NotLoadedProject` with `MarkTestProject` for simplification
- Introduced `MarkTestProject` class to streamline test setup.
- Removed redundant inline implementations in test classes.
- Improved code maintainability and readability.
