+++ test/src/org/omegat/gui/editor/mark/ComesFromAutoTMMarkerTest.java
@@ -38,63 +38,18 @@ import org.junit.Test;
 import org.omegat.core.Core;
 import org.omegat.core.TestCoreInitializer;
 import org.omegat.core.data.EntryKey;
-import org.omegat.core.data.NotLoadedProject;
-import org.omegat.core.data.ProjectProperties;
-import org.omegat.core.data.ProjectTMX;
 import org.omegat.core.data.SourceTextEntry;
-import org.omegat.core.data.TMXEntry;
 import org.omegat.core.segmentation.SRX;
 import org.omegat.core.segmentation.Segmenter;
-import org.omegat.util.Language;
 
 public class ComesFromAutoTMMarkerTest extends MarkerTestBase {
-
-    private ProjectTMX projectTMX;
 
     @Before
     public void preUp() throws Exception {
         TestCoreInitializer.initEditor(editor);
-        Core.setSegmenter(new Segmenter(SRX.getDefault()));
-        projectTMX = new ProjectTMX();
-        projectTMX.load(new Language("en"), new Language("fr"), true,
-                    Paths.get("test/data/autotmx/auto1.tmx").toFile(), Core.getSegmenter());
-        Core.setProject(new NotLoadedProject() {
-            @Override
-            public boolean isProjectLoaded() {
-                return true;
-            }
-
-            @Override
-            public ProjectProperties getProjectProperties() {
-                return new ProjectProperties() {
-                    @Override
-                    public Language getSourceLanguage() {
-                        return new Language("en");
-                    }
-
-                    @Override
-                    public Language getTargetLanguage() {
-                        return new Language("fr");
-                    }
-                };
-            }
-
-            @Override
-            public TMXEntry getTranslationInfo(SourceTextEntry ste) {
-                if (ste == null || projectTMX == null) {
-                    return EMPTY_TRANSLATION;
-                }
-                TMXEntry r = projectTMX.getMultipleTranslation(ste.getKey());
-                if (r == null) {
-                    r = projectTMX.getDefaultTranslation(ste.getSrcText());
-                }
-                if (r == null) {
-                    r = EMPTY_TRANSLATION;
-                }
-                return r;
-            }
-
-        });
+        Segmenter segmenter = new Segmenter(SRX.getDefault());
+        Core.setSegmenter(segmenter);
+        Core.setProject(new MarkTestProject(Paths.get("test/data/autotmx/auto1.tmx").toFile(), segmenter));
     }
 
     @Test
