+++ test/src/org/omegat/gui/editor/mark/ComesFromMTMarkerTest.java
@@ -40,76 +40,24 @@ import org.omegat.core.Core;
 import org.omegat.core.TestCoreInitializer;
 import org.omegat.core.data.DataUtils;
 import org.omegat.core.data.EntryKey;
-import org.omegat.core.data.NotLoadedProject;
 import org.omegat.core.data.PrepareTMXEntry;
-import org.omegat.core.data.ProjectProperties;
-import org.omegat.core.data.ProjectTMX;
 import org.omegat.core.data.SourceTextEntry;
-import org.omegat.core.data.TMXEntry;
 import org.omegat.core.matching.NearString;
 import org.omegat.core.segmentation.SRX;
 import org.omegat.core.segmentation.Segmenter;
-import org.omegat.util.Language;
 
 public class ComesFromMTMarkerTest extends MarkerTestBase {
 
-    private ProjectTMX projectTMX;
-    private Path tmroot;
     private Path project;
 
     @Before
     public void preUp() throws Exception {
-        tmroot = Paths.get("test/data/tmx/");
+        Path tmroot = Paths.get("test/data/tmx/");
         project = tmroot.resolve("mt/mt1.tmx"); // should under 'mt'
         TestCoreInitializer.initEditor(editor);
         Segmenter segmenter =  new Segmenter(SRX.getDefault());
         Core.setSegmenter(segmenter);
-        projectTMX = new ProjectTMX();
-        projectTMX.load(new Language("en"), new Language("fr"), true,
-                project.toFile(), segmenter);
-        Core.setProject(new NotLoadedProject() {
-            @Override
-            public boolean isProjectLoaded() {
-                return true;
-            }
-
-            @Override
-            public ProjectProperties getProjectProperties() {
-                return new ProjectProperties() {
-                    @Override
-                    public Language getSourceLanguage() {
-                        return new Language("en");
-                    }
-
-                    @Override
-                    public Language getTargetLanguage() {
-                        return new Language("fr");
-                    }
-
-                    @Override
-                    public String getTMRoot() {
-                        return tmroot.toString();
-                    }
-
-                };
-            }
-
-            @Override
-            public TMXEntry getTranslationInfo(SourceTextEntry ste) {
-                if (ste == null || projectTMX == null) {
-                    return EMPTY_TRANSLATION;
-                }
-                TMXEntry r = projectTMX.getMultipleTranslation(ste.getKey());
-                if (r == null) {
-                    r = projectTMX.getDefaultTranslation(ste.getSrcText());
-                }
-                if (r == null) {
-                    r = EMPTY_TRANSLATION;
-                }
-                return r;
-            }
-
-        });
+        Core.setProject(new MarkTestProject(project.toFile(), segmenter));
     }
 
     @Test
